+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/EdcSubmodelFacade.java
@@ -23,6 +23,8 @@
  ********************************************************************************/
 package org.eclipse.tractusx.irs.edc.client;
 
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
 import lombok.RequiredArgsConstructor;
@@ -80,22 +82,10 @@ public class EdcSubmodelFacade {
             throw new EdcClientException(cause);
         }
     }
-
-    @SuppressWarnings("PMD.PreserveStackTrace")
-    public EndpointDataReference getEndpointReferenceForAsset(final String endpointAddress, final String filterKey,
-            final String filterValue) throws EdcClientException {
-        try {
-            return client.getEndpointReferenceForAsset(endpointAddress, filterKey, filterValue).get();
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            return null;
-        } catch (ExecutionException e) {
-            final Throwable cause = e.getCause();
-            if (cause instanceof EdcClientException exceptionCause) {
-                throw exceptionCause;
-            }
-            throw new EdcClientException(cause);
-        }
+    
+    public List<CompletableFuture<EndpointDataReference>> getEndpointReferencesForAsset(final String endpointAddress,
+            final String filterKey, final String filterValue) throws EdcClientException {
+        return client.getEndpointReferencesForAsset(endpointAddress, filterKey, filterValue);
     }
 
 }